\name{acc.ggplot2-package}
\alias{acc.ggplot2-package}
\alias{acc.ggplot2}
\docType{package}
\title{
Fuctions to Extend ggplot2
}
\description{
A (eventual) collection of tools to extend and speed up coding for repeated uses of plotting functions that use ggplot2. 
}
\details{
\tabular{ll}{
Package: \tab acc.ggplot2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-06\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}